ai-appointment-chatbot/backend/requirements.txt
Flask==2.1.1
SQLAlchemy==1.4.27
psycopg2-binary==2.9.1
redis==4.0.2
requests==2.26.0
python-dotenv==0.19.2
twilio==6.62.0
sendgrid==6.9.1
alembic==1.7.5
openai==0.11.3
fastapi==0.68.1
uvicorn==0.15.0